Use relative include dir to fix subdir usage
To make it possible to use IrrlichtMt with `add_subdirectory()` command
use a relative include path instead of using `CMAKE_SOURCE_DIR`. When
using IrrlichtMt as a subdirectory the `CMAKE_SOURCE_DIR` variable points
to the root of the Minetest repo instead of the root of the IrrlichtMt
source directory, which breaks the build because the `IrrCompileConfig.h`
include can't be found.