determine compiler architecture with basename
The mingw build script for CI would not work with absolute paths for the compiler, because of a check for 32bit vs 64 bit architectures that assumed the compiler to be invoked as a command instead of a path to the compiler. This commit fixes the issue by checking the basename of the given compiler, so the check now treats paths the same way it used to treat commands.