use version range up to 3.9
The OLD behavior of a CMake policy is deprecated by default, so a version range can be used to set all policies to NEW on the appropriate versions. It is difficult to test this extensively, but it is unlikely to cause a problem, and could fix some bugs for people using newer versions. I picked 3.9 as the top of the version range to minimize the effects, because 3.9 will set CMP0069 to NEW, which is the focus of this commit.