add unit test framework to build
The CMake build will now support CTest. There are two known caveats: libIrrlichtMt.a has to have been linked to libdl (I have no idea why) and when executed with ctests, the tests do not show debugging information that is desplayed when simply running the executable. There is one test added, the disambiguateTextures test, and this one was chosen because it validates the working directory.