fix unit test warnings
I do not know that the pedantic flag is very important, and removing this comma to comply with it will probably be disagreeable to some people. My reason for complying with the test is that it is part of the coding style precedented by the previous authors of the library, and if warnings are ignored, warnings that actually indicate a bug are easy to miss.