fix overload hiding
Overloading a method in a subclass hides the definitions of the superclass. I don't think this is causing an issue because the subclass explicitly overrides both of the superclass methods, but now there is no question. There are some vptr errors ubsan complains about which this may be related to, but I do not know.

handle missing enumeration values in switch

ESNRP_LIGHT and ESNRP_SHADOW were not handled in the scene node registration. They are now explicitly ignored.

remove extraenous semicolons

always have defaults in color converter switch

The defaults were enabled only in release. Rationale for changing this is that the behavior is the same whether or not default is explicit. So disabling defaults for debugging is not really helpful, except that it displays a warning.

fix root cause of stringop warning

After researching the cause of the warning thoroughly, I realized that GCC is noticing the array[length] = 0 anti-pattern, where it expects array[length - 1] = 0. Sure enough, setting an intermediate variable to hold the allocation length (including the null terminator) lets us use the standard idom, which GCC has no qualms with.