Pass CFLAGS and CPPFLAGS to build process
When overrideing cmpiler options in cmake CMake does not by default honour
enviornment set compiler options, this has to be manually considered.
In Debian, CFLAGS and CPPFLAGS are injected by the enviornmnent to
set some archive-wide defaults, for example enabling hardening features.

(If the environment variables are not set, this patch has no effect.)

Patch origin:
https://salsa.debian.org/games-team/minetest/-/blob/master/debian/patches/enable_hardening.patch
(minus some extra blank lines, which I missed when creating the patch for Debian))