Specify dependencies on object library targets
The minimum CMake version was updated to 3.12 in #270. This cleans up
the build using the feature to use target_link_libraries on object
library targets that was added in CMake 3.12.

It has been tested on Linux with OpenGL. I will test it with more
settings and on more systems as the work progresses.